Fixes pLabel double scan number bug
pLabel export now can handle multiple identification assignments to the same spectrum.
It now takes all MGF titles and assigns them sequentially to the matching combination of rawfile name, scan number and precursor charge.
If the titles in the xTable and the MGF file do not match, pLabel raises an exception